fixing tests

+++ Tests/Test10-Learning.cpp
@@ -81,8 +81,8 @@ protected:
         }
         distrReal = dynamic_cast<EFG::strct::QueryHandler*>(this->referenceModel.get())->getMarginalDistribution((*vars.rbegin())->name());
         distrLearnt = dynamic_cast<EFG::strct::QueryHandler*>(this->trainedModel.get())->getMarginalDistribution((*vars.rbegin())->name());
-        EXPECT_LE(fabsf(distrLearnt.front() - distrReal.front()), 0.1f);
-        EXPECT_LE(fabsf(distrLearnt.back() - distrReal.back()), 0.1f);
+        EXPECT_LE(fabsf(static_cast<float>(distrLearnt.front() - distrReal.front())), 0.1f);
+        EXPECT_LE(fabsf(static_cast<float>(distrLearnt.back() - distrReal.back())), 0.1f);
         // check tuned values
         auto finalWeight = this->trainedModel->getWeights();
         {
